pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jamguard"
version = "0.1.0"
description = "Offline GNSS array analysis toolkit for beamforming and anti-jam experiments"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "JamGuard Capstone Team" }]
dependencies = [
  "numpy>=1.26",
  "scipy>=1.12",
  "matplotlib>=3.8",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
]

[project.scripts]
jamguard = "jamguard.cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_configs = true
